changes service signature of delete, update & add fixed individual component fixes-add/ update/delete

+++ src/main/java/com/interglobe/todo/controller/ItemController.java
@@ -1,13 +1,10 @@
 package com.interglobe.todo.controller;
 
-import java.net.URI;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-
-import org.springframework.http.ResponseEntity;
-
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,8 +14,6 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
 
 import com.interglobe.todo.domain.Item;
 import com.interglobe.todo.exception.ItemNotFoundException;
@@ -53,30 +48,37 @@ public class ItemController {
 	}
 	
 	@PostMapping("/")
-	public ResponseEntity<Object> addItem(@RequestBody Item item) {
+	public List<Item> addItem(@RequestBody Item item) {
 		if(item.getId() != null) {
 			throw new ItemNotFoundException("Invalid input - Item Id Not passed");
 		}
-		Item savedItem = repository.save(item);
-		repository.findById(savedItem.getId());
-		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
-		        .buildAndExpand(savedItem.getId()).toUri();
-		return ResponseEntity.created(location).build();
-	}
-	
-	@DeleteMapping("/id/{id}")
-	public void deleteItem(@PathVariable Integer id) {
-		repository.deleteById(id);
+		item.setPending(false);
+		repository.save(item);
+		return repository.findByIsPending(false);
 	}
 	
 	@PutMapping("/id/{id}")
-	public ResponseEntity<Object> updateItem(@RequestBody Item item, @PathVariable Integer id) {
+	public List<List<Item>> updateItem(@RequestBody Item item, @PathVariable Integer id) {
 		if(!repository.findById(id).isPresent()) {
 			throw new ItemNotFoundException("Item Id "+ id +"Not found");
 		}
 		item.setId(id);
-		Item updatedItem = repository.save(item);
-		repository.findById(updatedItem.getId());
-		return ResponseEntity.noContent().build();
+		repository.save(item);
+		List<List<Item>> updatedItemsLists = new ArrayList<>();
+		List<Item> pendingList = repository.findByIsPending(false);
+		List<Item> completedList = repository.findByIsPending(true);
+		updatedItemsLists.add(0, pendingList);
+		updatedItemsLists.add(1, completedList);
+		return updatedItemsLists;
 	}
+	
+	@DeleteMapping("/type/{type}/id/{id}")
+	public List<Item> deleteItem(@PathVariable boolean type, @PathVariable Integer id) {
+		if(repository.findByIsPending(type).isEmpty()) {
+			throw new ItemNotFoundException("No Item found for Type "+type);
+		}
+		repository.deleteById(id);
+		return repository.findByIsPending(type);
+	}
+	
 }
